improve random ID, add BeforeMapping into the CharacterMapper

+++ src/main/java/mate/academy/rickandmorty/dto/external/PageInfoDto.java
@@ -1,12 +1,11 @@
 package mate.academy.rickandmorty.dto.external;
 
-import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class PageInfoDto {
-    private BigDecimal count;
-    private BigDecimal pages;
+    private Integer count;
+    private Integer pages;
     private String next;
     private String prev;
 }
