+++ src/main/java/mate/academy/rickandmorty/dto/internal/CharacterShortInfoDto.java
@@ -1,9 +1,7 @@
 package mate.academy.rickandmorty.dto.internal;
 
-import java.math.BigDecimal;
-
 public record CharacterShortInfoDto(
-        BigDecimal id,
+        Long id,
         String externalId,
         String name,
         String status,
